Enhance documentation for EdgeX Foundry
This update improves explanations and descriptions related to the EdgeX Foundry framework and its architecture. It better explains how application services help manage sensor data in IoT solutions. The role and operation of the BI application gateway is now illustrated in a new dedicated file. Furthermore, overall documentation clarity is increased by restructuring the sections and headings, and additional links for further reading are also included.